+++ Runtime/ScriptableObjects/AnimationCurveReference.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace Jeomseon.ScriptableObjects
+namespace Jeomseon.Unity.EditorToolkit.ScriptableObjects
 {
     [CreateAssetMenu(fileName = "AnimcaitonCurve", menuName = "Scriptable Object/Animation Curve")]
     public sealed class AnimationCurveReference : UnityEngine.ScriptableObject